chore: common util for v3
